changes to personal todo
