+++ src/config/constants.config.ts
@@ -1,2 +1,1 @@
-export const PriorityEnum = ['Low', 'Medium', 'High'];
 export const SALT_ROUNDS = 10;
