+++ src/controllers/task.controller.ts
@@ -1,7 +1,6 @@
 import { Request, Response } from "express";
 import { Task } from "../models/Task.model";
-import { User } from "../models/User.model";
-
+// Create Personal Task
 const createPersonalTaskController = async (req: Request, res: Response) => {
     const task = await Task.create({
         title: req.body.title,
@@ -15,11 +14,13 @@ const createPersonalTaskController = async (req: Request, res: Response) => {
         },
     });
 };
-
+// Fetch Personal Task
 const getPersonalTaskController = async (req: Request, res: Response) => {
     const tasks = await Task.find();
-    res.json({ status: 200, message: "Success", data: tasks })
+    const filteredTasks = tasks.filter((x) => x?.userId.toString() === req?.body?.user?._id.toString())
+    res.json({ status: 200, message: "Success", data: filteredTasks })
 }
+// Update Personal Task
 const updatePersonalTaskController = async (req: Request, res: Response) => {
     const taskId = req?.params?.taskId;
     const task = await Task.findOneAndUpdate({ _id: taskId }, {
@@ -39,6 +40,7 @@ const updatePersonalTaskController = async (req: Request, res: Response) => {
         }
     })
 }
+// Delete Personal Task
 const deletePersonalTaskController = async (req: Request, res: Response) => {
     const taskId = req?.params?.taskId;
     await Task.findOneAndDelete({ _id: taskId });
