+++ src/index.ts
@@ -1,8 +1,8 @@
 import cors from "cors";
 import dotenv from "dotenv";
+dotenv.config();
 import express, { Express } from "express";
 import { rootRouter } from "./routes/index.route";
-dotenv.config();
 
 import connectDB from "./config/db.config";
 connectDB();
