+++ src/typings/types.ts
@@ -1,7 +1,17 @@
+import mongoose from "mongoose";
+import { PriorityEnum } from "../typings/enum";
+
 export type ProviderUserProfile = {
   sub: string;
   name: string;
   email: string;
   username?: string;
   id?: string;
+};
+export type PersonalTask = {
+  title: string;
+  description?: string;
+  tags?: string[];
+  priority?: PriorityEnum;
+  userId: mongoose.Types.ObjectId;
 };
